requirements.txt
azure-search-documents==11.6.0b7
azure-core>=1.30.0
openai>=1.30.1
python-dotenv>=1.0.1
pypdf>=4.2.0
tiktoken>=0.7.0
streamlit>=1.35.0
